Results controller. Added more functionality to repo and service. Custom JQL query fixed. Tested with swagger.

+++ src/main/java/com/dam/kanpeki/model/custom/ResultPerCategoryData.java
@@ -18,7 +18,7 @@ public class ResultPerCategoryData implements Serializable {
 
 	private Long categoryId;
 
-	private double numResults;
+	private Long numResults;
 
 	private double avgResults;
